Made program dialog its own component
